refactor: extract hooks from Monaco component
