+++ src/components/molecules/Monaco/Monaco.styled.tsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const HiddenInputContainer = styled.div`
+  width: 0;
+  height: 0;
+  overflow: hidden;
+`;
+
+export const HiddenInput = styled.input`
+  opacity: 0;
+`;
+
+export const MonacoContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  padding-left: 0px;
+  padding-right: 8px;
+  margin: 0px;
+  margin-bottom: 20px;
+`;
