+++ src/components/molecules/Monaco/useDebouncedCodeSave.ts
@@ -1,0 +1,59 @@
+import {useDebounce} from 'react-use';
+import {monaco} from 'react-monaco-editor';
+import {useAppDispatch} from '@redux/hooks';
+import {logMessage} from '@redux/services/log';
+import {updateFileEntry, updateResource} from '@redux/reducers/main';
+import {ResourceMapType} from '@models/appstate';
+
+function useDebouncedCodeSave(
+  editor: monaco.editor.IStandaloneCodeEditor | null,
+  orgCode: string,
+  code: string,
+  isDirty: boolean,
+  isValid: boolean,
+  resourceMap: ResourceMapType,
+  selectedResourceId: string | undefined,
+  selectedPath: string | undefined,
+  setOrgCode: (newOrgCode: string) => void
+) {
+  const dispatch = useAppDispatch();
+  const saveContent = () => {
+    let value = null;
+    if (editor) {
+      value = editor.getValue();
+    } else {
+      return;
+    }
+    // is a file and no resource selected?
+    if (selectedPath && !selectedResourceId) {
+      try {
+        dispatch(updateFileEntry({path: selectedPath, content: value}));
+        setOrgCode(value);
+      } catch (e) {
+        logMessage(`Failed to update file ${e}`, dispatch);
+      }
+    } else if (selectedResourceId && resourceMap[selectedResourceId]) {
+      try {
+        dispatch(updateResource({resourceId: selectedResourceId, content: value.toString()}));
+        setOrgCode(value);
+      } catch (e) {
+        logMessage(`Failed to update resource ${e}`, dispatch);
+      }
+    }
+  };
+
+  useDebounce(
+    () => {
+      if (!isDirty || !isValid) {
+        return;
+      }
+      if (orgCode !== undefined && code !== undefined && orgCode !== code) {
+        saveContent();
+      }
+    },
+    500,
+    [code]
+  );
+}
+
+export default useDebouncedCodeSave;
